+++ src/services/schema-export/tests/models/all/kotlin/ChildEmbeddedType.kt
@@ -1,0 +1,12 @@
+// Please note : @LinkingObjects and default values are not represented in the schema and thus will not be part of the generated models
+package your.package.name.here
+
+import io.realm.RealmObject
+import io.realm.annotations.RealmClass
+
+@RealmClass(embedded = true)
+open class ChildEmbeddedType : RealmObject() {
+
+    var id: Long = 0
+
+}