+++ src/services/schema-export/tests/models/all/kotlin/RequiredTypes.kt
@@ -3,6 +3,8 @@ package your.package.name.here
 
 import io.realm.RealmObject
 import java.util.Date
+import org.bson.types.ObjectId
+import org.bson.types.Decimal128
 import io.realm.RealmList
 import io.realm.annotations.Required
 
@@ -15,6 +17,8 @@ open class RequiredTypes : RealmObject() {
     var stringRequired: String = ""
     var dateRequired: Date = Date()
     var dataRequired: ByteArray = ByteArray(0)
+    var objectIdRequired: ObjectId = ObjectId()
+    var decimal128Required: Decimal128 = Decimal128()
     @Required
     var boolRequiredArray: RealmList<Boolean> = RealmList()
     @Required
@@ -29,6 +33,10 @@ open class RequiredTypes : RealmObject() {
     var dateRequiredArray: RealmList<Date> = RealmList()
     @Required
     var dataRequiredArray: RealmList<ByteArray> = RealmList()
+    @Required
+    var objectIdRequiredArray: RealmList<ObjectId> = RealmList()
+    @Required
+    var decimal128RequiredArray: RealmList<Decimal128> = RealmList()
     var objectRequiredArray: RealmList<RequiredTypes> = RealmList()
 
 }