+++ src/ui/RealmBrowser/Content/CreateObjectDialog/types/DefaultControl.tsx
@@ -20,13 +20,20 @@ import React from 'react';
 import { Alert } from 'reactstrap';
 import Realm from 'realm';
 
+export type AlertColorType = 'info' | 'warning' | 'danger';
+
 export interface IDefaultControlProps {
   property: Realm.ObjectSchemaProperty;
   message?: string;
+  color?: AlertColorType;
 }
 
-export const DefaultControl = ({ property, message }: IDefaultControlProps) => (
-  <Alert color="warning">
+export const DefaultControl = ({
+  property,
+  message,
+  color = 'warning',
+}: IDefaultControlProps) => (
+  <Alert color={color}>
     {message ?? `Cannot select "${property.type}" yet`}
   </Alert>
 );