+++ src/ui/RealmBrowser/Content/CreateObjectDialog/types/ListControl.tsx
@@ -24,6 +24,7 @@ import Realm from 'realm';
 import { isPrimitive } from '../../../primitives';
 
 import { ITypeControlProps, TypeControl } from './TypeControl';
+import { DefaultControl } from './DefaultControl';
 
 interface IItemProps
   extends Pick<
@@ -146,8 +147,17 @@ export const ListControl = ({
 }: ITypeControlProps): React.ReactElement<ITypeControlProps> => {
   if (!property.objectType) {
     return <Alert color="danger">Expected an objectType</Alert>;
+  } else if (isEmbeddedType(property.objectType)) {
+    return (
+      <DefaultControl
+        color="info"
+        property={property}
+        message="Embedded objects can be created from the table, once this object has been created."
+      />
+    );
   } else {
     const itemProperty = getItemProperty(property);
+
     return (
       <section className="CreateObjectDialog__ListControl">
         <List